fix: let command line arguments take precedence over scoped config
This a change that most won't notice. Nonetheless it's breaking.

This CLI program knows a concept of scoped config.
e.g. when you sign into a different API endpoint
(staging, enterprise) the token is saved scoped to the API.
This way the CLI can store tokens for multiple endpoints
and it's not necessary to re-login whenever the endpoint is
changed.

There was a bug so that a scoped config would take
precedence over CLI arguments. This is now fixed.

BREAKING CHANGE: Before the scoped admin and token took precedence over command line flags.