feat: send reminder comment to stale initial PRs
Turns out some of our users are confused why Greenkeeper is not enabled
for their repos and it turns out, it's because they never merged the
initial PR. With this, we send out a reminder comment on the PR so
people have a clearer understanding of what happens.

We don't nag them again once we reminded them for that repo. We also
don't comment on very old PRs, as this likely causes more problems for
our users then it does good (super outdated dependencies etc.)

Hope this helps a couple of users to run
Greenkeeper :)