feat: retry 404 errors on `repos.get()`
adds logging so we can find out why we are getting a 404 here. If the
case is that we are getting a notification about this before we actually
do have access, we retry with the usual 5/3000 policy.

If we can’t get a repo info in five tries, we ignore the error and log
the result.