fix(push): report malformed greenkeeper.json error on GitHub push
This one required a refactor of the GitHub push job, so it is a
little bit more involved.

Before, getGreenkeeperConfigFile was swallowing exceptions for
getting the file from GitHub and for parsing it.

Now, a parse error throws an error that in turn causes the
invalid-config-file job as per #678 and #672.

Also, GitHub errors now cause log entries, and return an empty
object, like the earlier default behaviour.