feat: support multiple exec servers
A list of active exec servers is kept in CouchDB. For each exec
server request, we get that list.

Each job increments an active jobs counter, this allows us to
determine te server with the fewest jobs to send a new one to.

On network errors, we now try again with another server. Retries
are implemented with recursive function calls, and capped at 4
retries.

500 errors are not handled specially any more, as the exec server
mostly hides them now.