fix(monorepo): avoid fetching infos for internal dependencies
yarn monorepos are identified by packageJson.workspaceRoot.
if we are handling one of those, we might encounter a [sub-]
package.json that specifies a dependency like this:
"@reponame/dependency": "*" which looks like a scoped npm
dependency, but is in reality a in-monorepo dependency.
For those dependencies, we should not try to fetch infos
from npm, because the dependencies are not there.